requirements.txt
# --- Core ML / NLP ---
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.7.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# --- Text Extraction ---
pdfplumber>=0.10.0
python-docx>=1.0.0
pytesseract>=0.3.10
Pillow>=10.0.0

# --- NLP Preprocessing ---
spacy>=3.7.0

# --- Vector Search ---
faiss-cpu>=1.7.4

# --- Backend API ---
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.5.0

# --- Frontend ---
streamlit>=1.32.0
plotly>=5.18.0

# --- Utilities ---
tqdm>=4.66.0
matplotlib>=3.7.0
seaborn>=0.13.0
joblib>=1.3.0
python-dotenv>=1.0.0
